_ugc_api/requirements.txt
aiokafka==0.8.0
fastapi==0.95.0
gunicorn==20.1.0
pydantic==1.10.6
PyJWT==2.6.0
uvicorn==0.21.1

# Sub. dependencies
anyio==3.6.2
async-timeout==4.0.2
click==8.1.3
h11==0.14.0
idna==3.4
kafka-python==2.0.2
packaging==23.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
